Testing finally understood, by default Json.NET does nothing on serialization for naming
Testing finally understood, by default Json.NET does nothing on serialization for naming. This means looking back at my trouble with serialization on a different project, I must have had a spelling mistake, plain and simple. There is no other explanation. However, I now have some pretty powerful tools for verifying these kinds of things. I will expand upon the serialization aspects on custom classes and types next to allow for recursive serialization of lists and nested objects, and being able to recursively parse to specified depths, as does Json.NET. This is somewhat fun for me to do, so even though you may think I am insane, I only increase my understanding of how things work, and have experience building such libraries. I want to be able to verify serialization cases for nested objects and may next test Microsoft Json serializers and have a testbed for learning aspects of the libraries quickly in ways others can not. Currently, this is a limitation on the Json.NET Extensions methods I created here for recursive parsing on nested objects, but I was testing for simpler explanations for where I thought in the following scenario: when Json.NET serialization settings were initialized with NullValueHandling.Ignore on a global sort of level, rather than individual property attributes, and I removed all but the JsonProperty attributes NullValueHandling attributes, minus leaving the PropertyName changes to the models, that it was ignoring the property name changes. However, based on my testing here, I have identified this was not the case, and validated that I must have just missed the simplest of simplest of things, like I did not have uppercase ID on the end of a MemberID, but MemberId, even in the Json.NET Property Attribute for PropertyName, and I just was not catching it every time I reviewed the code :)... However, I lost the original code to go back and verify it again, and face other Network issues regressing my ability to test directly in the original code base in one of many branches, and for whatever reason I cannot find that code, and I must have disgarded the changes on the other project. However, based on the amount of extention library creation and testing I just performmed here to verify my sanity after the fact, I am very comfortable with how the Json.NET settings perform, and I got backup in future projects of how these operations happen, and whether or not I can get away with certain attribute naming paradigms and be able to backup my decisions and errors with more firepower of knowing 100% how they work together, and that I had a simple error that caused me much confusion and grief, to the point I would write an entire test extension library to verify my findings without access to the original source mistake. -Ryan Mauldin